RequestService
*

Creating and implementing basic methods for request service

*

+++ src/main/java/br/com/vsgdev/emotionsRoute/repository/RequestRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.vsgdev.emotionsRoute.model.Request;
 
 @Repository
-public interface RequestRepository extends CrudRepository<Request, Integer> {
+public interface RequestRepository extends CrudRepository<Request, Long> {
 
 }
